Need to re-think how we are doing things
After discussing the analysis code with Dennis, a few of these
changes need to be re-thought. We agreed that the method of
including the simulation-specific settings in the output ROOT file
(in the TRexGeant4, simulation it's done by including the
TRexSettings custom class in the ROOT file) is the smart way of
doing things.