Adding try/catch statements for out_of_range error
Looks like the detector numbering for Griffin has been changed
somehow - as it currently is, out_of_range errors are produced when
either the detectorID = 16 or the crystalID = 4. In the NTuple code,
vectors for the detectorID and crystalID run from 0-15 and 0-3
respectively, however the tree produced by the Griffinv10
simulation run from 0-16 and 0-4. Still needs to be resolved.