Adding more histograms
Some problems with the analysis when comparing the histograms produced by my version of the analysis vs. that done in the old simulation w/ ReconstructSim. The first discrepancy is when looking at the originXY histogram. We are also sampling from a circle for the XY distribion, this could be related.